Atualização

+++ Bank-Client/Project-Bank-main/src/main/java/com/bank/messasing/ClientConsumer.java
@@ -1,13 +1,12 @@
 package com.bank.messasing;
 
-import java.awt.List;
 
 import org.springframework.amqp.rabbit.annotation.RabbitListener;
+import org.springframework.amqp.rabbit.core.RabbitTemplate;
 import org.springframework.stereotype.Service;
 
 import com.bank.config.RabbitMQConfig;
 import com.bank.dto.ClientRequestDTO;
-import com.bank.exception.ResourceNotFoundException;
 import com.bank.model.Client;
 import com.bank.repository.ClientRepository;
 
@@ -15,35 +14,51 @@ import com.bank.repository.ClientRepository;
 public class ClientConsumer {
 
 	private final ClientRepository clientRepository;
+	private final RabbitTemplate rabbitTemplate;
 
-	public ClientConsumer(ClientRepository clientRepository) {
+	public ClientConsumer(ClientRepository clientRepository, RabbitTemplate rabbitTemplate) {
 
 		this.clientRepository = clientRepository;
-	}
-	
-	@RabbitListener(queues = RabbitMQConfig.QUEUE_CLIENT_DELETE)
-	public void delete(ClientRequestDTO dto) {
-		Client client = clientRepository.findById(dto.getId()).orElseThrow(()-> new ResourceNotFoundException("Client not found"));
-		clientRepository.delete(client);
-		System.out.println("Sucess Deleted RabbitMQ");
-	}
-	
-	@RabbitListener(queues = RabbitMQConfig.QUEUE_CLIENT_UPDATE)
-	public void update(ClientRequestDTO updateDTO) {
-	Client client = clientRepository.findById(updateDTO.getId()).orElseThrow(() -> new ResourceNotFoundException("Client not found"));
-	client.setName(updateDTO.getName());
-	client.setAge(updateDTO.getAge());
-	client.setAddress(updateDTO.getAddress());
-	client.setEmail(updateDTO.getEmail());
-	client.setSalary(updateDTO.getSalary());
-	
-	clientRepository.save(client);
-	System.out.println("Sucess Update RabbitMQ!");
-	
-	
+		this.rabbitTemplate = rabbitTemplate;
+
 	}
 
-	//Salvar
+	// Delete - completo com tratamento de erro de loop infinity
+	@RabbitListener(queues = RabbitMQConfig.QUEUE_CLIENT_DELETE)
+	public void delete(ClientRequestDTO dto) {
+		Client client = clientRepository.findById(dto.getId()).orElse(null);
+
+		if (client != null) {
+			clientRepository.delete(client);
+			System.out.println("Success deleted RabbitMQ id: " + dto.getId());
+		} else {
+			rabbitTemplate.convertAndSend(RabbitMQConfig.DLQ_CLIENT_DELETE, dto);
+			System.out.println("Client not found, sent to DLQ DELETE RabbitMQ id: " + dto.getId());
+		}
+
+	}
+
+	// Update - completo com tratamento de erro de loop infinity
+	@RabbitListener(queues = RabbitMQConfig.QUEUE_CLIENT_UPDATE)
+	public void update(ClientRequestDTO updateDTO) {
+		Client client = clientRepository.findById(updateDTO.getId()).orElse(null);
+
+		if (client != null) {
+			client.setName(updateDTO.getName());
+			client.setAge(updateDTO.getAge());
+			client.setAddress(updateDTO.getAddress());
+			client.setEmail(updateDTO.getEmail());
+			client.setSalary(updateDTO.getSalary());
+			clientRepository.save(client);
+			System.out.println("Success update RabbitMQ id: "+updateDTO.getId());
+		} else {
+			rabbitTemplate.convertAndSend(RabbitMQConfig.DLQ_CLIENT_UPDATE, updateDTO);
+			System.out.println("Client not found, sent to DLQ UPDATE RabbitMQ id: " + updateDTO.getId());
+		}
+
+	}
+
+	// Salvar
 	@RabbitListener(queues = RabbitMQConfig.QUEUE_CLIENT_SAVE)
 	public void save(ClientRequestDTO request) {
 		Client client = new Client();
@@ -57,6 +72,4 @@ public class ClientConsumer {
 		System.out.println("Client Created!");
 	}
 
-	
-	
 }
